+++ MapboxCoreNavigation/MMEEventsManager.swift
@@ -332,12 +332,12 @@ extension RouteLegProgress {
         get {
             return [
                 "upcomingInstruction": upComingStep?.instructions ?? NSNull(),
-                "upcomingType": upComingStep?.maneuverType?.description ?? NSNull(),
-                "upcomingModifier": upComingStep?.maneuverDirection?.description ?? NSNull(),
+                "upcomingType": upComingStep?.maneuverType.description ?? NSNull(),
+                "upcomingModifier": upComingStep?.maneuverDirection.description ?? NSNull(),
                 "upcomingName": upComingStep?.names?.joined(separator: ";") ?? NSNull(),
                 "previousInstruction": currentStep.instructions,
-                "previousType": currentStep.maneuverType?.description ?? NSNull(),
-                "previousModifier": currentStep.maneuverDirection?.description ?? NSNull(),
+                "previousType": currentStep.maneuverType.description,
+                "previousModifier": currentStep.maneuverDirection.description,
                 "previousName": currentStep.names?.joined(separator: ";") ?? NSNull(),
                 "distance": Int(currentStep.distance),
                 "duration": Int(currentStep.expectedTravelTime),