+++ MapboxNavigation/ManeuverView.swift
@@ -67,9 +67,9 @@ public class ManeuverView: UIView {
         }
         
         var flip: Bool = false
-        let type: ManeuverType = step.maneuverType ?? .turn
+        let type: ManeuverType = step.maneuverType != .none ? step.maneuverType : .turn
         let angle = ((step.finalHeading ?? 0) - (step.initialHeading ?? 0)).wrap(min: -180, max: 180)
-        let direction: ManeuverDirection = step.maneuverDirection ?? ManeuverDirection(angle: Int(angle))
+        let direction: ManeuverDirection = step.maneuverDirection != .none ? step.maneuverDirection : ManeuverDirection(angle: Int(angle))
 
         switch type {
         case .merge: