+++ MapboxNavigationTests/InstructionsBannerViewTests.swift
@@ -28,7 +28,7 @@ class InstructionsBannerViewTests: FBSnapshotTestCase {
         super.setUp()
         recordMode = false
         
-        let instruction = VisualInstructionComponent(text: nil, imageURL: shieldURL)
+        let instruction = VisualInstructionComponent(type: .destination, text: nil, imageURL: shieldURL)
         let shieldKey = instruction.shieldKey()
         SDImageCache.shared().store(shieldImage, forKey: shieldKey)
     }
@@ -45,9 +45,9 @@ class InstructionsBannerViewTests: FBSnapshotTestCase {
         view.distance = 482
         
         let instructions = [
-            VisualInstructionComponent(text: "US 45", imageURL: nil),
-            VisualInstructionComponent(text: "/", imageURL: nil),
-            VisualInstructionComponent(text: "Chicago", imageURL: nil)
+            VisualInstructionComponent(type: .destination, text: "US 45", imageURL: nil),
+            VisualInstructionComponent(type: .destination, text: "/", imageURL: nil),
+            VisualInstructionComponent(type: .destination, text: "Chicago", imageURL: nil)
         ]
         
         view.set(instructions, secondaryInstruction: nil)
@@ -63,8 +63,8 @@ class InstructionsBannerViewTests: FBSnapshotTestCase {
         view.distance = 482
 
         let instructions = [
-            VisualInstructionComponent(text: "I 280", imageURL: shieldURL),
-            VisualInstructionComponent(text: "US 45 / Chicago / US 45 / Chicago", imageURL: nil)
+            VisualInstructionComponent(type: .destination, text: "I 280", imageURL: shieldURL),
+            VisualInstructionComponent(type: .destination, text: "US 45 / Chicago / US 45 / Chicago", imageURL: nil)
         ]
 
         view.set(instructions, secondaryInstruction: nil)
@@ -80,10 +80,10 @@ class InstructionsBannerViewTests: FBSnapshotTestCase {
         view.distance = 482
         
         let primary = [
-            VisualInstructionComponent(text: "I 280", imageURL: shieldURL),
-            VisualInstructionComponent(text: "South", imageURL: nil)
+            VisualInstructionComponent(type: .destination, text: "I 280", imageURL: shieldURL),
+            VisualInstructionComponent(type: .destination, text: "South", imageURL: nil)
         ]
-        let secondary = [VisualInstructionComponent(text: "US 45 / Chicago", imageURL: nil)]
+        let secondary = [VisualInstructionComponent(type: .destination, text: "US 45 / Chicago", imageURL: nil)]
         
         view.set(primary, secondaryInstruction: secondary)
 
@@ -98,9 +98,9 @@ class InstructionsBannerViewTests: FBSnapshotTestCase {
         view.distance = 482
         
         let primary = [
-            VisualInstructionComponent(text: "I 280", imageURL: shieldURL)
+            VisualInstructionComponent(type: .destination, text: "I 280", imageURL: shieldURL)
         ]
-        let secondary = [VisualInstructionComponent(text: "Mountain View Test", imageURL: nil)]
+        let secondary = [VisualInstructionComponent(type: .destination, text: "Mountain View Test", imageURL: nil)]
 
         view.set(primary, secondaryInstruction: secondary)
         
@@ -122,15 +122,15 @@ class InstructionsBannerViewTests: FBSnapshotTestCase {
         instructionsBannerView.distance = 482
         
         let primary = [
-            VisualInstructionComponent(text: "I 280", imageURL: shieldURL)
+            VisualInstructionComponent(type: .destination, text: "I 280", imageURL: shieldURL)
         ]
-        let secondary = [VisualInstructionComponent(text: "US 45 / Chicago", imageURL: nil)]
+        let secondary = [VisualInstructionComponent(type: .destination, text: "US 45 / Chicago", imageURL: nil)]
         
         instructionsBannerView.set(primary, secondaryInstruction: secondary)
 
         
         let primaryThen = [
-            VisualInstructionComponent(text: "I 280", imageURL: shieldURL)
+            VisualInstructionComponent(type: .destination, text: "I 280", imageURL: shieldURL)
         ]
         
         nextBannerView.instructionLabel.instruction = primaryThen