+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changes to the Mapbox Navigation SDK for iOS
 
-## master
+## 0.13.0
+
+* Removed `RouteControllerDelegate.routeController(_:shouldIncrementLegWhenArrivingAtWaypoint:)` and `NavigationViewControllerDelegate.navigationViewController(_:shouldIncrementLegWhenArrivingAtWaypoint:)`. `RouteControllerDelegate.routeController(_:didArriveAt:)` and `NavigationViewControllerDelegate.navigationViewController(_:didArriveAt:)` now return a Boolean that determines whether the route controller automatically advances to the next leg of the route. (#1038)
+* Fixed an issue where `NavigationViewControllerDelegate.navigationViewController(_:didArriveAt:)` was called twice at the end of the route. (#1038)
+* Added a Danish localization. (#1031)
+* Added a partial Hebrew localization. (#1043)
 
 ## v0.12.2 (January 12, 2018)